requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.43
python-telegram-bot==20.7
python-dotenv==1.0.0
Werkzeug==3.0.1
requests==2.31.0
gunicorn==21.2.0
